ge_datastore_accessor_agent/requirements.txt
google-adk
google-auth
requests

# Mandatory for Agent Engine
google-cloud-aiplatform[agent_engines,adk]
